Make .indices a data property. Closes #29.
Making .indices a caching getter in the spec makes the function identity
of the getter observable. This makes it necessary for implementations to
create a new Function object per result object of
`RegExp.prototype.exec`. The intended laziness of .indices is possible
in implementations even with it as a data property in spec.